% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTodosAddin.R
\name{findTodosAddin}
\alias{findTodosAddin}
\title{Addin to Find todos}
\usage{
findTodosAddin()
}
\value{
function to run todoAddin
}
\description{
Addin to Find todos
}
\examples{
  todoAddin:::findTodosAddin()
}
